+++ integreat_cms/cms/views/events/event_list_view.py
@@ -8,10 +8,7 @@ from django.utils.decorators import method_decorator
 from django.utils.translation import ugettext as _
 from django.views.generic import TemplateView
 
-from ...constants import (
-    status,
-    translation_status,
-)
+from ...constants import translation_status
 from ...decorators import permission_required
 from ...forms import EventFilterForm
 from .event_context_mixin import EventContextMixin
@@ -33,12 +30,7 @@ class EventListView(TemplateView, EventContextMixin):
     #: Template for list of archived events
     template_archived = "events/event_list_archived.html"
     #: The context dict passed to the template (see :class:`~django.views.generic.base.ContextMixin`)
-    extra_context = {
-        "current_menu_item": "events",
-        "WEBAPP_URL": settings.WEBAPP_URL,
-        "PUBLIC": status.PUBLIC,
-        "DEEPL_ENABLED": settings.DEEPL_ENABLED,
-    }
+    extra_context = {"current_menu_item": "events"}
     #: Whether or not to show archived events
     archived = False