+++ integreat_cms/cms/views/pages/page_context_mixin.py
@@ -1,10 +1,9 @@
 import logging
 
-from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic.base import ContextMixin
 
-from ...constants import translation_status, status
+from ...constants import translation_status
 from ...utils.translation_utils import ugettext_many_lazy as __
 
 logger = logging.getLogger(__name__)
@@ -32,11 +31,6 @@ class PageContextMixin(ContextMixin):
             {
                 "current_menu_item": "pages",
                 "translation_status": translation_status,
-                "PUBLIC": status.PUBLIC,
-                "DRAFT": status.DRAFT,
-                "REVIEW": status.REVIEW,
-                "AUTO_SAVE": status.AUTO_SAVE,
-                "WEBAPP_URL": settings.WEBAPP_URL,
                 "archive_dialog_title": _(
                     "Please confirm that you really want to archive this page"
                 ),