+++ integreat_cms/core/context_processors.py
@@ -2,7 +2,9 @@
 Context processors pass additional variables to templates (see :ref:`context-processors`).
 """
 from django.conf import settings
-from integreat_cms import __version__
+
+from .. import __version__
+from ..cms.constants import status
 
 
 # pylint: disable=unused-variable
@@ -20,28 +22,38 @@ def version_processor(request):
 
 
 # pylint: disable=unused-variable
-def push_notification_processor(request):
+def settings_processor(request):
     """
-    This context processor injects the setting :attr:`~integreat_cms.core.settings.FCM_ENABLED` into the template context.
+    This context processor injects a few settings into the template context.
 
     :param request: The current http request
     :type request: ~django.http.HttpRequest
 
-    :return: The currently installed version of this package
+    :return: A few of our settings
     :rtype: dict
     """
-    return {"FCM_ENABLED": settings.FCM_ENABLED}
+    return {
+        "FCM_ENABLED": settings.FCM_ENABLED,
+        "BRANDING": settings.BRANDING,
+        "WEBAPP_URL": settings.WEBAPP_URL,
+        "DEEPL_ENABLED": settings.DEEPL_ENABLED,
+    }
 
 
 # pylint: disable=unused-variable
-def branding_processor(request):
+def constants_processor(request):
     """
-    This context processor injects the setting :attr:`~integreat_cms.core.settings.BRANDING` into the template context.
+    This context processor injects some of our constants into the template context.
 
     :param request: The current http request
     :type request: ~django.http.HttpRequest
 
-    :return: The currently installed version of this package
+    :return: A few of our constants
     :rtype: dict
     """
-    return {"BRANDING": settings.BRANDING}
+    return {
+        "DRAFT": status.DRAFT,
+        "REVIEW": status.REVIEW,
+        "AUTO_SAVE": status.AUTO_SAVE,
+        "PUBLIC": status.PUBLIC,
+    }